More fully fleshed-out version of pickling test harness.

+++ patsy/test_pickling.py
@@ -1,7 +1,10 @@
+from __future__ import print_function
+
 import six
 from six.moves import cPickle as pickle
 
 import os
+import shutil
 
 from patsy import EvalFactor
 
@@ -10,7 +13,6 @@ PICKE_TESTCASES_ROOTDIR = os.path.join(os.path.dirname(__file__), '..', 'pickle_
 pickling_testcases = {
     "evalfactor_simple": EvalFactor("a+b"),
     }
-
 
 def test_pickling_same_version_roundtrips():
     for obj in six.itervalues(pickling_testcases):
@@ -39,22 +41,30 @@ def test_unpickling_future_gives_sensible_error_msg():
     # TODO How do we do this? And how do we test it then?
     pass
 
-def save_pickle_testcases(version):
+def create_pickles(version):
     pickle_testcases_dir = os.path.join(PICKE_TESTCASES_ROOTDIR, version)
-    # Fails if the directory already exists, which is what we want here
-    # since we don't want to overwrite testcases accidentally.
-    os.mkdir(pickle_testcases_dir)
+    if os.path.exists(pickle_testcases_dir):
+        raise OSError("{} already exists. Aborting.".format(pickle_testcases_dir))
+    pickle_testcases_tempdir = pickle_testcases_dir + "_inprogress"
+    os.mkdir(pickle_testcases_tempdir)
 
-    for name, obj in six.iteritems(pickling_testcases):
-        with open(os.path.join(pickle_testcases_dir, '{}.pickle'.format(name)), 'wb') as f:
-            pickle.dump(obj, f, pickle.HIGHEST_PROTOCOL)
+    try:
+        for name, obj in six.iteritems(pickling_testcases):
+            with open(os.path.join(pickle_testcases_tempdir, "{}.pickle".format(name)), "wb") as f:
+                pickle.dump(obj, f, pickle.HIGHEST_PROTOCOL)
+    except Exception:
+        print("Exception during creation of pickles for {}. Removing directory.".format(version))
+        shutil.rmtree(pickle_testcases_tempdir)
+        raise
+    finally:
+        os.rename(pickle_testcases_tempdir, pickle_testcases_dir)
+    print("Successfully created pickle testcases for new version {}.".format(version))
 
 if __name__ == "__main__":
     import argparse
 
-    # Should we use a "create-pickles" sub-command to make things clearer?
-    arg_parser = argparse.ArgumentParser()
+    arg_parser = argparse.ArgumentParser(description="Create and save pickle testcases for a new version of patsy.")
     arg_parser.add_argument("version", help="The version of patsy for which to save a new set of pickle testcases.")
     args = arg_parser.parse_args()
 
-    save_pickle_testcases(args.version)
+    create_pickles(args.version)
